fix: improve lock when using the CLI to build
